Encode sequences as bitfield in u8
This commit changes the encoding of a single nucleotide from a [bool; 5] to a
u8 with the 5 lower bits being set like the 5 bools previously.
This causes sequences to be represented in 5x less space, which should translate
to better generated code and higher speeds, as well as smaller databases.

Note that the database representation have now changed, since that depends on the
internal representation of the windows.
Also note that the distance computation now in `get_distances` is now 1 per
different nucleotide, instead of 2. Hence, the divisions by 2 have been removed.